requirements.txt
beautifulsoup4==4.13.3
flasgger==0.9.7.1
Flask==3.1.0
flask_sqlalchemy==3.1.1
openpyxl==3.1.5
Requests==2.32.3
